refactor(engine): cleanup taken flows on activating
Taking a sequence flow contained some complexity in
BpmnStateTransitionBehavior, because it needed to keep track of the
taken sequence flows, as well as taking the sequence flow. Taking the
sequence flow would result in cleanup of knowledge of which flows were
taken, so it needed to start tracking it already before actually taking
the sequence flow.

The cleanup was only performed when the target parallel gateway is
actually activated. So, this cleanup could also be moved to the
activation of the parallel gateway.
