+++ engine/src/main/java/io/zeebe/engine/state/appliers/ProcessInstanceSequenceFlowTakenApplier.java
@@ -7,17 +7,14 @@
  */
 package io.zeebe.engine.state.appliers;
 
-import io.zeebe.engine.processing.deployment.model.element.ExecutableFlowNode;
 import io.zeebe.engine.processing.deployment.model.element.ExecutableSequenceFlow;
 import io.zeebe.engine.state.TypedEventApplier;
-import io.zeebe.engine.state.instance.IndexedRecord;
 import io.zeebe.engine.state.instance.StoredRecord.Purpose;
 import io.zeebe.engine.state.mutable.MutableElementInstanceState;
 import io.zeebe.engine.state.mutable.MutableProcessState;
 import io.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import io.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.zeebe.protocol.record.value.BpmnElementType;
-import java.util.stream.Collectors;
 
 /** Applies state changes for `ProcessInstance:Sequence_Flow_Taken` */
 final class ProcessInstanceSequenceFlowTakenApplier
@@ -48,55 +45,24 @@ final class ProcessInstanceSequenceFlowTakenApplier
     flowScopeInstance.incrementActiveSequenceFlows();
     elementInstanceState.updateInstance(flowScopeInstance);
 
-    final var process = processState.getProcessByKey(value.getProcessDefinitionKey());
-    final var element =
-        process.getProcess().getElementById(value.getElementId(), ExecutableSequenceFlow.class);
-    final var target = element.getTarget();
+    final var sequenceFlow =
+        processState.getFlowElement(
+            value.getProcessDefinitionKey(),
+            value.getElementIdBuffer(),
+            ExecutableSequenceFlow.class);
+    final var target = sequenceFlow.getTarget();
 
     if (target.getElementType() != BpmnElementType.PARALLEL_GATEWAY) {
       return;
     }
 
     // store which sequence flows are taken as deferred records
+    // these are cleaned up on ParallelGateway:Element_Activating
     elementInstanceState.storeRecord(
         key,
         value.getFlowScopeKey(),
         value,
         ProcessInstanceIntent.SEQUENCE_FLOW_TAKEN,
         Purpose.DEFERRED);
-
-    final var tokensBySequenceFlow =
-        elementInstanceState.getDeferredRecords(value.getFlowScopeKey()).stream()
-            .filter(record -> record.getState() == ProcessInstanceIntent.SEQUENCE_FLOW_TAKEN)
-            .filter(record -> isIncomingSequenceFlow(record, target))
-            .collect(Collectors.groupingBy(record -> record.getValue().getElementIdBuffer()));
-
-    if (tokensBySequenceFlow.size() == target.getIncoming().size()) {
-      // all incoming sequence flows have been taken, so the gateway is going to be activated
-
-      // cleanup deferred records
-      tokensBySequenceFlow.forEach(
-          (sequenceFlow, tokens) -> {
-            final var firstToken = tokens.get(0);
-            elementInstanceState.removeStoredRecord(
-                value.getFlowScopeKey(), firstToken.getKey(), Purpose.DEFERRED);
-          });
-
-      // finally write the active token changes to state
-      elementInstanceState.updateInstance(flowScopeInstance);
-    }
-  }
-
-  // copied from BpmnStateTransitionBehavior
-  private boolean isIncomingSequenceFlow(
-      final IndexedRecord record, final ExecutableFlowNode parallelGateway) {
-    final var elementId = record.getValue().getElementIdBuffer();
-
-    for (final ExecutableSequenceFlow incomingSequenceFlow : parallelGateway.getIncoming()) {
-      if (elementId.equals(incomingSequenceFlow.getId())) {
-        return true;
-      }
-    }
-    return false;
   }
 }
