--- util/src/main/java/io/zeebe/util/collection/ListUtil.java
@@ -1,20 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Zeebe Community License 1.1. You may not use this file
- * except in compliance with the Zeebe Community License 1.1.
- */
-package io.zeebe.util.collection;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-public final class ListUtil {
-
-  @SafeVarargs
-  public static <T> List<T> concat(final List<T>... lists) {
-    return Stream.of(lists).flatMap(List::stream).collect(Collectors.toList());
-  }
-}
